PayPal: Remove unnecessary paypalCountry
The new variable paypalCountry is not needed. Let's use the part after
the dash in the languageCode instead. Add a country to the few
languageCodes not having it and where it's not the same as the language
part.

Even if a languageCode is set to only "de" it will still work given it's
the code for both the country and the language.